Move RichGroup and RichCommand out of rich_click.py
